% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_slider.R
\name{commit_slider_ui}
\alias{commit_slider_ui}
\title{Commit Range Slider Module}
\usage{
commit_slider_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
A custom dual-handle slider for selecting commit ranges
}
