% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_issue_ids}
\alias{fix_issue_ids}
\title{Fix ID serialization issues for a single Issue object}
\usage{
fix_issue_ids(issue)
}
\arguments{
\item{issue}{Single issue object with potential floating point ID fields}
}
\value{
Issue object with IDs converted to character strings
}
\description{
Converts floating point IDs to character strings to prevent serialization issues
when passing Issue objects to Rust functions.
}
