% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_milestones}
\alias{get_milestones}
\title{Get milestones with ID fix applied}
\usage{
get_milestones(working_dir)
}
\arguments{
\item{working_dir}{Working directory path}
}
\value{
List of milestones with fixed IDs
}
\description{
Wrapper around get_milestones_extr that automatically fixes floating
point ID serialization issues by converting IDs to character strings.
}
