% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_multiple_milestone_issues}
\alias{get_multiple_milestone_issues}
\title{Get multiple milestone issues with ID fix applied}
\usage{
get_multiple_milestone_issues(milestones, working_dir)
}
\arguments{
\item{milestones}{List of milestone objects}

\item{working_dir}{Working directory path}
}
\value{
Named list of milestone issues with fixed IDs
}
\description{
Wrapper around get_multiple_milestone_issues_impl that automatically fixes floating
point ID serialization issues by converting IDs to character strings.
}
