% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{ghqc_config_setup}
\alias{ghqc_config_setup}
\title{Setup GHQC Configuration Repository}
\usage{
ghqc_config_setup(config_url = NULL, config_path = NULL)
}
\arguments{
\item{config_url}{Character. URL of the configuration repository to clone.}

\item{config_path}{Character. Path where the configuration should be stored.}
}
\value{
Character message indicating setup status.
}
\description{
Set-up the configuration repository. If \code{config_url} is provided, the repository will be cloned to \verb{$XDG_DATA_HOME/ghqc/config}
If not AND \code{GHQC_CONFIG_URL} is provided, the repository will be cloned to \verb{$XDG_DATA_HOME/ghqc/<repo name>}
}
