% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{ghqc_config_status}
\alias{ghqc_config_status}
\title{Check GHQC Configuration Status}
\usage{
ghqc_config_status(config_path = NULL)
}
\arguments{
\item{config_path}{Character. Path to the configuration directory. If NULL,
uses default configuration location.}
}
\value{
Prints configuration status message.
}
\description{
Provides the status of the configuration repository found at \code{config_path} (default according to rules in \code{ghqc_config_setup()})
}
