% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_slider.R
\name{commit_slider_ui}
\alias{commit_slider_ui}
\title{Commit Range Slider Module}
\usage{
commit_slider_ui(id)
}
\arguments{
\item{id}{Module ID}

\item{commits}{Data frame with columns: hash, message, qc_class, edits_file}

\item{default_from}{Default "from" commit (short SHA)}

\item{default_to}{Default "to" commit (short SHA)}
}
\description{
A custom dual-handle slider for selecting commit ranges
}
